Reordered data, added Slitherdrake
I changed the order of the data to be in line with the order it appears in the Rostrum. I also updated all Drakewatcher Manuscripts as of 10.1.7 PTR; I commented out some items that have no known source in the live game (including PTR data). I also added a file for the Winding Slitherdrake.